Share instructor epics fetch logic
